backend/requirements.txt
boto3==1.26.22
botocore==1.29.22
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-decouple==3.6
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
